requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
langchain
streamlit
